SETI-2840: Do not delete duplicate jobs in one pipeline (job tree)
Do not delete duplicates in one job tree. Having duplicate job
in a job_tree is a viable use case. Even if a job is defined
multiple times in a tree, it is only launched once. The result
is saved and reused if needed.

Another problem is the way duplicates are deleted: only duplicates
which are direct children are deleted (indirect children are left
there). This inconsistent behavior is not nice.

Modify JobTree.getJobs to return a set of jobs instead of a list,
so it does not have a problem with duplicate jobs in the tree.