SETI-2840: Introduce job sets into pipelines' job trees
Current tree structure of jobs in a pipeline do not allow to define
a job which depends on successful execution of multiple jobs.

Modify JobTree so that a single node can contain either a single
job (same as before) or a set of jobs. Children of this node are
executed only after all jobs in this node finish successfully.

YAML cannot handle sets natively and an ordinary python set cannot
be a dictionary key (because it is not hashable). Let's introduce
a "!jobset" tag which translates into a python frozenset object,
which is hashable.

A job set in zuul layout is then defined as follows:
!jobset [job1, job2, job3]